moving parameters.py to automates/utils/
